requirements.txt
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-groq
authlib
google-cloud-firestore
google-auth
tabulate